+++ packages/pyspartalib/tests/pyspartalib/script/project/test_base_pipeline.py
@@ -8,7 +8,7 @@ from pyspartalib.context.custom.type_context import Type
 from pyspartalib.context.default.string_context import Strs
 from pyspartalib.context.extension.path_context import PathPair
 from pyspartalib.script.path.modify.get_resource import get_resource
-from pyspartalib.script.project.base_pipeline import BasePipeline
+from pyspartalib.script.pipeline.base_pipeline import BasePipeline
 
 
 def _difference_error(result: Type, expected: Type) -> None:
